HTML entity encoding before POST
Functionality to encode HTML entities for a POST, need this if data is
an object with properties whose values contains strings with quotation
marks.
HTML entity encoding is done using the htmlentities() function from the
phpjs.org project.
Note that some browsers will POST the string htmlentity-encoded whilst
others will decode it before POSTing.
It is recommended that on the server, htmlentity decoding is done
irrespective.