Allow to conditionally disable android POST restriction
The way I am using your script doesn't seem to have any issues with android, so I wanted to have the option to disable the prohibition of downloading from a POST request in android.  I've been testing with a few 4.+ versions with post and they all download the files properly and I am able to view them without any issues.